--- a/charset_normalizer/constant.py
+++ b/charset_normalizer/constant.py
@@ -1,12 +1,12 @@
 from codecs import BOM_UTF8, BOM_UTF16_BE, BOM_UTF16_LE, BOM_UTF32_BE, BOM_UTF32_LE
 from encodings.aliases import aliases
 from re import IGNORECASE, compile as re_compile
-from typing import Dict, List, Set, Union
+from typing import Dict, List, Set, Union, Final
 
 from .assets import FREQUENCIES
 
 # Contain for each eligible encoding a list of/item bytes SIG/BOM
-ENCODING_MARKS: Dict[str, Union[bytes, List[bytes]]] = {
+ENCODING_MARKS: Final[Dict[str, Union[bytes, List[bytes]]]] = {
     "utf_8": BOM_UTF8,
     "utf_7": [
         b"\x2b\x2f\x76\x38",
@@ -20,12 +20,12 @@ ENCODING_MARKS: Dict[str, Union[bytes, List[bytes]]] = {
     "utf_16": [BOM_UTF16_BE, BOM_UTF16_LE],
 }
 
-TOO_SMALL_SEQUENCE: int = 32
-TOO_BIG_SEQUENCE: int = int(10e6)
+TOO_SMALL_SEQUENCE: Final[int] = 32
+TOO_BIG_SEQUENCE: Final[int] = int(10e6)
 
-UTF8_MAXIMAL_ALLOCATION: int = 1112064
+UTF8_MAXIMAL_ALLOCATION: Final[int] = 1112064
 
-UNICODE_RANGES_COMBINED: Dict[str, range] = {
+UNICODE_RANGES_COMBINED: Final[Dict[str, range]] = {
     "Control character": range(31 + 1),
     "Basic Latin": range(32, 127 + 1),
     "Latin-1 Supplement": range(128, 255 + 1),
@@ -308,7 +308,7 @@ UNICODE_RANGES_COMBINED: Dict[str, range] = {
 }
 
 
-UNICODE_SECONDARY_RANGE_KEYWORD: List[str] = [
+UNICODE_SECONDARY_RANGE_KEYWORD: Final[List[str]] = [
     "Supplement",
     "Extended",
     "Extensions",
@@ -326,12 +326,12 @@ UNICODE_SECONDARY_RANGE_KEYWORD: List[str] = [
     "Tags",
 ]
 
-RE_POSSIBLE_ENCODING_INDICATION = re_compile(
+RE_POSSIBLE_ENCODING_INDICATION: Final = re_compile(
     r"(?:(?:encoding)|(?:charset)|(?:coding))(?:[\:= ]{1,10})(?:[\"\']?)([a-zA-Z0-9\-_]+)(?:[\"\']?)",
     IGNORECASE,
 )
 
-IANA_SUPPORTED: List[str] = sorted(
+IANA_SUPPORTED: Final[List[str]] = sorted(
     filter(
         lambda x: x.endswith("_codec") is False
         and x not in {"rot_13", "tactis", "mbcs"},
@@ -339,10 +339,10 @@ IANA_SUPPORTED: List[str] = sorted(
     )
 )
 
-IANA_SUPPORTED_COUNT: int = len(IANA_SUPPORTED)
+IANA_SUPPORTED_COUNT: Final[int] = len(IANA_SUPPORTED)
 
 # pre-computed code page that are similar using the function cp_similarity.
-IANA_SUPPORTED_SIMILAR: Dict[str, List[str]] = {
+IANA_SUPPORTED_SIMILAR: Final[Dict[str, List[str]]] = {
     "cp037": ["cp1026", "cp1140", "cp273", "cp500"],
     "cp1026": ["cp037", "cp1140", "cp273", "cp500"],
     "cp1125": ["cp866"],
@@ -431,7 +431,7 @@ IANA_SUPPORTED_SIMILAR: Dict[str, List[str]] = {
 }
 
 
-CHARDET_CORRESPONDENCE: Dict[str, str] = {
+CHARDET_CORRESPONDENCE: Final[Dict[str, str]] = {
     "iso2022_kr": "ISO-2022-KR",
     "iso2022_jp": "ISO-2022-JP",
     "euc_kr": "EUC-KR",
@@ -467,7 +467,7 @@ CHARDET_CORRESPONDENCE: Dict[str, str] = {
 }
 
 
-COMMON_SAFE_ASCII_CHARACTERS: Set[str] = {
+COMMON_SAFE_ASCII_CHARACTERS: Final[Set[str]] = {
     "<",
     ">",
     "=",
@@ -486,10 +486,10 @@ COMMON_SAFE_ASCII_CHARACTERS: Set[str] = {
 }
 
 
-KO_NAMES: Set[str] = {"johab", "cp949", "euc_kr"}
-ZH_NAMES: Set[str] = {"big5", "cp950", "big5hkscs", "hz"}
+KO_NAMES: Final[Set[str]] = {"johab", "cp949", "euc_kr"}
+ZH_NAMES: Final[Set[str]] = {"big5", "cp950", "big5hkscs", "hz"}
 
-LANGUAGE_SUPPORTED_COUNT: int = len(FREQUENCIES)
+LANGUAGE_SUPPORTED_COUNT: Final[int] = len(FREQUENCIES)
 
 # Logging LEVEL bellow DEBUG
-TRACE: int = 5
+TRACE: Final[int] = 5