--- a/charset_normalizer/constant.py
+++ b/charset_normalizer/constant.py
@@ -2,7 +2,7 @@ import sys
 from codecs import BOM_UTF8, BOM_UTF16_BE, BOM_UTF16_LE, BOM_UTF32_BE, BOM_UTF32_LE
 from encodings.aliases import aliases
 from re import IGNORECASE, compile as re_compile
-from typing import Dict, List, Pattern, Set, Union
+from typing import Dict, List, Set, Union
 
 if sys.version_info <= (3, 8):
     from typing_extensions import Final
@@ -332,7 +332,7 @@ UNICODE_SECONDARY_RANGE_KEYWORD: Final[List[str]] = [
     "Tags",
 ]
 
-RE_POSSIBLE_ENCODING_INDICATION: Final[Pattern[str]] = re_compile(
+RE_POSSIBLE_ENCODING_INDICATION: Final = re_compile(
     r"(?:(?:encoding)|(?:charset)|(?:coding))(?:[\:= ]{1,10})(?:[\"\']?)([a-zA-Z0-9\-_]+)(?:[\"\']?)",
     IGNORECASE,
 )