--- a/dev-requirements.txt
+++ b/dev-requirements.txt
@@ -8,3 +8,4 @@ black==22.6.0
 flake8==5.0.4
 mypy==0.971
 isort
+typing_extensions